UserService is now used by TokenService; User feature doesn't use Token feature but vice versa
