+++ backend/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 # General
-server.servlet.context-path=/api/v1
+server.servlet.context-path=/api
 
 # Database configuration
 spring.datasource.url=jdbc:postgresql://treenity-db:5432/postgres
